+++ src/main/java/org/example/delivery/store/model/response/GetStoresResponse.java
@@ -1,4 +1,4 @@
-package org.example.delivery.store.model.dto.response;
+package org.example.delivery.store.model.response;
 
 import java.sql.Time;
 import org.example.delivery.common.domain.entity.Store;
