+++ src/test/java/org/example/delivery/unit/store/service/StoreServiceTest.java
@@ -15,9 +15,9 @@ import org.example.delivery.common.exception.base.BusinessException;
 import org.example.delivery.common.exception.base.ConflictException;
 import org.example.delivery.common.exception.base.NotFoundException;
 import org.example.delivery.menu.repository.MenuRepository;
-import org.example.delivery.store.model.dto.request.StoreRequest;
-import org.example.delivery.store.model.dto.response.GetStoreByIdResponse;
-import org.example.delivery.store.model.dto.response.StoreResponse;
+import org.example.delivery.store.model.request.StoreRequest;
+import org.example.delivery.store.model.response.GetStoreByIdResponse;
+import org.example.delivery.store.model.response.StoreResponse;
 import org.example.delivery.store.repository.StoreRepository;
 import org.example.delivery.auth.repository.UserRepository;
 import org.example.delivery.store.service.StoreService;
